fix: duplicate new windows when clicking topic links
